+++ container-engine-lib/lib/backend_impls/kubernetes/kubernetes_kurtosis_backend_api_container_functions.go
@@ -245,6 +245,7 @@ func (backend *KubernetesKurtosisBackend) CreateAPIContainer(
 				consts.PodExecsKubernetesResource,
 				consts.ServicesKubernetesResource,
 				consts.PersistentVolumeClaimsKubernetesResource,
+				consts.JobsKubernetesResource,
 			},
 		},
 		{